+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## xx.x.x
+* Added a config method to disable user location tracking: `setDisableLocation`
+* Added a call to disable user location tracking: `disableLocation:`
+
+* Updated underlying Android SDK version to 23.8.4
+* Underlying iOS SDK version is 23.8.3
+
+## xx.x.x-np
+* Added a config method to disable user location tracking: `setDisableLocation`
+* Added a call to disable user location tracking: `disableLocation:`
+
+* Updated underlying Android SDK version to 23.8.4
+* Underlying iOS SDK version is 23.8.3
+
 ## 23.8.4
 * Added a call to enroll users to A/B experiment with experiment ID : `testingEnrollIntoABExperiment:`
 * Added a call to exit users from A/B experiment with experiment ID : `testingExitABExperiment:`